+++ vision/datasets/download_mnist_test.go
@@ -1,4 +1,4 @@
-package vision
+package datasets
 
 import (
 	"io/ioutil"
